requirements.txt
azure-ai-vision-imageanalysis==1.0.0
azure-core==1.33.0
Pillow==10.0.0
colorama==0.4.6
arabic-reshaper==3.0.0
python-bidi==0.4.2
python-dotenv==1.0.0
boto3==1.34.0
sqlalchemy>=2.0
asyncpg
pydantic-settings>=2.10.1
python-multipart
uvicorn
pytest~=8.4.1
httpx~=0.28.1
pytest-asyncio

fastapi~=0.115.12
botocore~=1.34.162
pydantic~=2.11.5
starlette~=0.46.2
fastapi_mail
python-jose[cryptography]
passlib
bcrypt==4.0.1
email-validator
